Tiles and Hero canvas position now passed in to draw method

+++ GridCombat/Actors/Tile.cs
@@ -11,7 +11,7 @@
     {
         #region Fields
 
-        public static float diameter = 50;
+        public static int diameter = 50;
 
         #endregion
